+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## In development
+
+### Changed
+* Update Django to version 3.2.4 LTS (#153)
+
+### Fixed
+* Fixed the result being different depending on the hint_attr_value of
+  search_entries (#158)
+* Fixed a problem in the processing of import entry (#159)
+
 ## v3.1.0
 
 ### Added