+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fixed the result being different depending on the hint_attr_value of
   search_entries (#158)
 * Fixed a problem in the processing of import entry (#159)
+* Fixed `urls` to avoid some warnings (#174)
 * Fixed a bug that raises an exception at API handler of update entry (#186)
 
 ## v3.1.0